port check 039: Perform all Fontforge checks
(issue #1388)
