+++ Lib/fontbakery/checks.py
@@ -114,135 +114,6 @@ def check_file_is_named_canonically(fb, font_fname):
     return False
 
 
-def perform_all_fontforge_checks(fb, validation_state):
-  def ff_check(description, condition, err_msg, ok_msg):
-    import sha
-    m = sha.new()
-    m.update(description)
-    short_hash = m.hexdigest()[:8]
-    fb.new_check("039-{}".format(short_hash),
-                 "fontforge-check: {}".format(description))
-    if condition is False:
-      fb.error("fontforge-check: {}".format(err_msg))
-    else:
-      fb.ok("fontforge-check: {}".format(ok_msg))
-
-  ff_check("Contours are closed?",
-           bool(validation_state & 0x2) is False,
-           "Contours are not closed!",
-           "Contours are closed.")
-
-  ff_check("Contours do not intersect",
-           bool(validation_state & 0x4) is False,
-           "There are countour intersections!",
-           "Contours do not intersect.")
-
-  ff_check("Contours have correct directions",
-           bool(validation_state & 0x8) is False,
-           "Contours have incorrect directions!",
-           "Contours have correct directions.")
-
-  ff_check("References in the glyph haven't been flipped",
-           bool(validation_state & 0x10) is False,
-           "References in the glyph have been flipped!",
-           "References in the glyph haven't been flipped.")
-
-  ff_check("Glyphs have points at extremas",
-           bool(validation_state & 0x20) is False,
-           "Glyphs do not have points at extremas!",
-           "Glyphs have points at extremas.")
-
-  ff_check("Glyph names referred to from glyphs present in the font",
-           bool(validation_state & 0x40) is False,
-           "Glyph names referred to from glyphs"
-           " not present in the font!",
-           "Glyph names referred to from glyphs"
-           " present in the font.")
-
-  ff_check("Points (or control points) are not too far apart",
-           bool(validation_state & 0x40000) is False,
-           "Points (or control points) are too far apart!",
-           "Points (or control points) are not too far apart.")
-
-  ff_check("Not more than 1,500 points in any glyph"
-           " (a PostScript limit)",
-           bool(validation_state & 0x80) is False,
-           "There are glyphs with more than 1,500 points!"
-           "Exceeds a PostScript limit.",
-           "Not more than 1,500 points in any glyph"
-           " (a PostScript limit).")
-
-  ff_check("PostScript has a limit of 96 hints in glyphs",
-           bool(validation_state & 0x100) is False,
-           "Exceeds PostScript limit of 96 hints per glyph",
-           "Font respects PostScript limit of 96 hints per glyph")
-
-  ff_check("Font doesn't have invalid glyph names",
-           bool(validation_state & 0x200) is False,
-           "Font has invalid glyph names!",
-           "Font doesn't have invalid glyph names.")
-
-  ff_check("Glyphs have allowed numbers of points defined in maxp",
-           bool(validation_state & 0x400) is False,
-           "Glyphs exceed allowed numbers of points defined in maxp",
-           "Glyphs have allowed numbers of points defined in maxp.")
-
-  ff_check("Glyphs have allowed numbers of paths defined in maxp",
-           bool(validation_state & 0x800) is False,
-           "Glyphs exceed allowed numbers of paths defined in maxp!",
-           "Glyphs have allowed numbers of paths defined in maxp.")
-
-  ff_check("Composite glyphs have allowed numbers"
-           " of points defined in maxp?",
-           bool(validation_state & 0x1000) is False,
-           "Composite glyphs exceed allowed numbers"
-           " of points defined in maxp!",
-           "Composite glyphs have allowed numbers"
-           " of points defined in maxp.")
-
-  ff_check("Composite glyphs have allowed numbers"
-           " of paths defined in maxp",
-           bool(validation_state & 0x2000) is False,
-           "Composite glyphs exceed"
-           " allowed numbers of paths defined in maxp!",
-           "Composite glyphs have"
-           " allowed numbers of paths defined in maxp.")
-
-  ff_check("Glyphs instructions have valid lengths",
-           bool(validation_state & 0x4000) is False,
-           "Glyphs instructions have invalid lengths!",
-           "Glyphs instructions have valid lengths.")
-
-  ff_check("Points in glyphs are integer aligned",
-           bool(validation_state & 0x80000) is False,
-           "Points in glyphs are not integer aligned!",
-           "Points in glyphs are integer aligned.")
-
-  # According to the opentype spec, if a glyph contains an anchor point
-  # for one anchor class in a subtable, it must contain anchor points
-  # for all anchor classes in the subtable. Even it, logically,
-  # they do not apply and are unnecessary.
-  ff_check("Glyphs have all required anchors.",
-           bool(validation_state & 0x100000) is False,
-           "Glyphs do not have all required anchors!",
-           "Glyphs have all required anchors.")
-
-  ff_check("Glyph names are unique?",
-           bool(validation_state & 0x200000) is False,
-           "Glyph names are not unique!",
-           "Glyph names are unique.")
-
-  ff_check("Unicode code points are unique?",
-           bool(validation_state & 0x400000) is False,
-           "Unicode code points are not unique!",
-           "Unicode code points are unique.")
-
-  ff_check("Do hints overlap?",
-           bool(validation_state & 0x800000) is False,
-           "Hints should NOT overlap!",
-           "Hinds do not overlap.")
-
-
 def check_non_ASCII_chars_in_ASCII_only_NAME_table_entries(fb, font):
   fb.new_check("074", "Are there non-ASCII characters"
                       " in ASCII-only NAME table entries ?")
